updated all size_t to unsigned and added debugging verbosity

+++ source/aggregators/FrameMesh.hpp
@@ -82,9 +82,9 @@ class FrameMesh {
          * 
          * @param column_line The column line index.
          * @param floor The floor index.
-         * @return size_t The vertex ID.
+         * @return unsigned The vertex ID.
          */
-        size_t get_vertix_id(int column_line, int floor)
+        unsigned get_vertix_id(int column_line, int floor)
         {
             if (column_line >= 0 && column_line <= nbays && floor >= 0 && floor <= nfloors)
             {
@@ -102,16 +102,16 @@ class FrameMesh {
          * @brief Gets the vertex IDs for all vertices at a specific floor and all bays.
          * 
          * @param floor The floor at which to get the vertices - can be from 0 to \ref nfloors.
-         * @return std::set<size_t> A set of vertex IDs at the specified floor.
+         * @return std::set<unsigned> A set of vertex IDs at the specified floor.
          */
-        std::set<size_t> get_vertices_ids_at_floor(int floor)
+        std::set<unsigned> get_vertices_ids_at_floor(int floor)
         {
             if (floor < 0 || floor > nfloors)
             {
                 std::cout << "FrameMesh::get_vertices_ids_at_floor expects floor >= 0 and <= nfloors, but got (floor/nfloors) = (" << floor << "/" << nfloors << ")" << std::endl;
                 exit(1); 
             }
-            std::set<size_t> vertices;
+            std::set<unsigned> vertices;
             for (int column_line_i = 0; column_line_i <= nbays; ++column_line_i)
             {
                 vertices.insert(get_vertix_id(column_line_i, floor));
@@ -122,11 +122,11 @@ class FrameMesh {
         /**
          * @brief Gets the vertices ids for the entire model.
          * 
-         * @return std::set<size_t> vertices_ids a std::set of vertices that are the ids of all vertices in the model.
+         * @return std::set<unsigned> vertices_ids a std::set of vertices that are the ids of all vertices in the model.
          */
-        std::set<size_t> get_vertices_ids()
+        std::set<unsigned> get_vertices_ids()
         {
-            std::set<size_t> vertices_ids;
+            std::set<unsigned> vertices_ids;
             for (int floor_i = 0; floor_i <= nfloors; ++floor_i)
             {
                 for (int column_line_i = 0; column_line_i <= nbays; ++column_line_i)
@@ -141,9 +141,9 @@ class FrameMesh {
          * @brief returns the column node ids at a given floor for a certain column_line including vertices. 
          * @param column_line the column line which the column we want is at. Ranges from 0 to \ref nbays.
          * @param floor floor at which the column starts. Ranges from 0 to < \ref nfloors.
-         * @return std::set<size_t> nodes a std::set of node ids for a column.
+         * @return std::set<unsigned> nodes a std::set of node ids for a column.
          */
-        std::set<size_t> get_column_node_ids(int column_line, int floor)
+        std::set<unsigned> get_column_node_ids(int column_line, int floor)
         {
             if (column_line < 0 || column_line > nbays || floor < 0 || floor >= nfloors)
             {
@@ -152,10 +152,10 @@ class FrameMesh {
                           << "Valid column_line range: [0, " << nbays << "], valid floor range: [0, " << nfloors - 1 << "]" << std::endl;
                 exit(1);
             }
-            std::set<size_t> nodes;
+            std::set<unsigned> nodes;
 
-            size_t starting_id = 1 + column_line*nodes_per_column_line + floor*column_divisions + column_line*nodes_per_full_bay;
-            for (size_t i = 1; i < nodes_per_column + 1; ++i)
+            unsigned starting_id = 1 + column_line*nodes_per_column_line + floor*column_divisions + column_line*nodes_per_full_bay;
+            for (unsigned i = 1; i < nodes_per_column + 1; ++i)
             {
                 nodes.insert(starting_id + i);
             }
@@ -167,24 +167,24 @@ class FrameMesh {
          * @param bay bay in which the beam is. Ranges from 1 to \ref nbays.
          * @param floor floor at which the beam is. Ranges from 1 to \ref nfloors.
          * @param include_vertices boolean to include or exclude vertices of the beam.
-         * @return std::set<size_t> A set of node IDs representing the nodes of the beam.
+         * @return std::set<unsigned> A set of node IDs representing the nodes of the beam.
          */
-        std::set<size_t> get_beam_node_ids(int bay, int floor, bool include_vertices = false)
+        std::set<unsigned> get_beam_node_ids(int bay, int floor, bool include_vertices = false)
         {
             if ((bay < 1) || (floor < 1) || floor > nfloors || bay > nbays)
             {
                 std::cout << "FrameMesh::get_beam_node_ids expects bay in range [1, " << nbays << "] and floor in range [1, " << nfloors << "], but got bay = " << bay << " and floor = " << floor << std::endl;
                 exit(1);
             }
-            std::set<size_t> nodes;
+            std::set<unsigned> nodes;
             
             if (include_vertices)
             {
                 nodes.insert(get_vertix_id(bay - 1, floor));
             }
 
-            size_t starting_id = bay*nodes_per_column_line + (bay - 1)*nodes_per_full_bay + (floor - 1) * nodes_per_beam;
-            for (size_t i = 1; i < nodes_per_beam + 1; ++i)
+            unsigned starting_id = bay*nodes_per_column_line + (bay - 1)*nodes_per_full_bay + (floor - 1) * nodes_per_beam;
+            for (unsigned i = 1; i < nodes_per_beam + 1; ++i)
             {
                 nodes.insert(starting_id + i);
             }
@@ -200,19 +200,19 @@ class FrameMesh {
          * @brief Gets the column line node IDs for a specific floor.
          * 
          * @param column_line The column_line index. Ranges from 0 to \ref nbays.
-         * @return std::set<size_t> A set of node IDs for the column line at the specified floor.
+         * @return std::set<unsigned> A set of node IDs for the column line at the specified floor.
          */
-        std::set<size_t> get_column_line_node_ids(int column_line)
+        std::set<unsigned> get_column_line_node_ids(int column_line)
         {
             if (column_line < 0 || column_line > nbays)
             {
                 std::cout << "FrameMesh::get_column_line_node_ids expects column_line >= 0 and <= nbays, but got (column_line/nbays) = (" << column_line << "/" << nbays << ")" << std::endl;
                 exit(1);
             }
-            std::set<size_t> nodes;
-            std::set<size_t> column_node_ids;
+            std::set<unsigned> nodes;
+            std::set<unsigned> column_node_ids;
 
-            for (size_t floor_i = 0; floor_i < nfloors; ++floor_i)
+            for (unsigned floor_i = 0; floor_i < nfloors; ++floor_i)
             {   
                 nodes.insert(get_vertix_id(column_line, floor_i));
                 column_node_ids = get_column_node_ids(column_line, floor_i);
@@ -227,17 +227,17 @@ class FrameMesh {
          * 
          * @param floor the floor at which the beam is located. Ranges from 1 to \ref nfloors.
          * @param include_vertices boolean to include or exclude vertices of the beam.
-         * @return std::set<size_t> A set of node IDs for the beam line at the specified beam line.
+         * @return std::set<unsigned> A set of node IDs for the beam line at the specified beam line.
          */
-        std::set<size_t> get_beam_line_node_ids(int floor, bool include_vertices = false)
+        std::set<unsigned> get_beam_line_node_ids(int floor, bool include_vertices = false)
         {
             if (floor <= 0 || floor > nfloors)
             {
                 std::cout << "FrameMesh::get_beam_line_node_ids expects floor > 0 and <= nfloors, but got (floor/nfloors) = (" << floor << "/" << nfloors << ")" << std::endl;
                 exit(1); 
             }
-            std::set<size_t> nodes;
-            std::set<size_t> bay_node_ids;
+            std::set<unsigned> nodes;
+            std::set<unsigned> bay_node_ids;
 
             for (int bay_i = 1; bay_i <= nbays; ++bay_i)
             {
@@ -253,14 +253,14 @@ class FrameMesh {
          * This function generates a vector of pairs, where each pair consists of a node ID and its corresponding coordinates.
          * The nodes are created for each column line in the frame.
          * 
-         * @return std::vector<std::pair<size_t, coords>> A vector of pairs mapping node IDs to their coordinates for column nodes.
+         * @return std::vector<std::pair<unsigned, coords>> A vector of pairs mapping node IDs to their coordinates for column nodes.
          */
-        std::vector<std::pair<size_t, coords>> create_column_node_coords_pairs()
+        std::vector<std::pair<unsigned, coords>> create_column_node_coords_pairs()
         {
-            std::vector<std::pair<size_t, coords>> nodes_id_coords_pairs;
+            std::vector<std::pair<unsigned, coords>> nodes_id_coords_pairs;
             for (int column_line_i = 0; column_line_i <= nbays; ++column_line_i)
             {
-                std::set<size_t> node_ids = get_column_line_node_ids(column_line_i);
+                std::set<unsigned> node_ids = get_column_line_node_ids(column_line_i);
                 real x_0 = column_line_i * bay_length;
                 real y_0 = 0.0;
                 real x = x_0;
@@ -279,20 +279,20 @@ class FrameMesh {
          * This function generates a vector of pairs, where each pair consists of a node ID and its corresponding coordinates.
          * The nodes are created for each beam line in the frame, *excluding* the vertices as these are created by \ref create_column_node_coords_pairs.
          * 
-         * @return std::vector<std::pair<size_t, coords>> A vector of pairs mapping node IDs to their coordinates for beam nodes.
+         * @return std::vector<std::pair<unsigned, coords>> A vector of pairs mapping node IDs to their coordinates for beam nodes.
          */
-        std::vector<std::pair<size_t, coords>> create_beam_node_coords_pairs()
+        std::vector<std::pair<unsigned, coords>> create_beam_node_coords_pairs()
         {
-            std::vector<std::pair<size_t, coords>> nodes_id_coords_pairs;
-            std::set<size_t> vertices = get_vertices_ids();
+            std::vector<std::pair<unsigned, coords>> nodes_id_coords_pairs;
+            std::set<unsigned> vertices = get_vertices_ids();
             for (int floor_i = 1; floor_i <= nfloors; ++floor_i)
             {
-                std::set<size_t> node_ids = get_beam_line_node_ids(floor_i, true);
+                std::set<unsigned> node_ids = get_beam_line_node_ids(floor_i, true);
                 real x_0 = 0.0;
                 real y_0 = floor_i * floor_height;
                 real x = x_0;
                 real y = y_0;
-                for (size_t node_id : node_ids)
+                for (unsigned node_id : node_ids)
                 {
                     if (!vertices.count(node_id))
                     {
@@ -308,11 +308,11 @@ class FrameMesh {
          * 
          * @details This function combines the node coordinate pairs for both column and beam nodes, sorts them by node ID, and returns the result.
          * 
-         * @return std::vector<std::pair<size_t, coords>> A sorted vector of pairs mapping node IDs to their coordinates for the entire frame.
+         * @return std::vector<std::pair<unsigned, coords>> A sorted vector of pairs mapping node IDs to their coordinates for the entire frame.
          */
-        std::vector<std::pair<size_t, coords>> get_node_coords_pairs()
+        std::vector<std::pair<unsigned, coords>> get_node_coords_pairs()
         {
-           std::vector<std::pair<size_t, coords>> nodes_coords_vector, beam_nodes_coords_vector;
+           std::vector<std::pair<unsigned, coords>> nodes_coords_vector, beam_nodes_coords_vector;
 
            nodes_coords_vector = create_column_node_coords_pairs();
            beam_nodes_coords_vector = create_beam_node_coords_pairs();
@@ -333,38 +333,38 @@ class FrameMesh {
          * @details This function generates a vector of pairs, where each pair consists of an element ID and a vector of node IDs that the element connects to.
          * The elements are created for each bay and floor in the frame.
          * 
-         * @return std::vector<std::pair<size_t, std::vector<size_t>>> A vector of pairs mapping element IDs to their corresponding node IDs.
+         * @return std::vector<std::pair<unsigned, std::vector<unsigned>>> A vector of pairs mapping element IDs to their corresponding node IDs.
          */
-        std::vector<std::pair<size_t, std::vector<size_t>>> map_elements_to_nodes()
+        std::vector<std::pair<unsigned, std::vector<unsigned>>> map_elements_to_nodes()
         {
-            std::vector<std::pair<size_t, std::vector<size_t>>> elements_map;
+            std::vector<std::pair<unsigned, std::vector<unsigned>>> elements_map;
             elements_map.reserve(nbays * nfloors * beam_divisions + (nbays + 1)*nfloors*column_divisions);
-            size_t element_id = 0;
+            unsigned element_id = 0;
 
             for (int bay_i = 1; bay_i <= nbays; ++bay_i) {
                 // first column line of each bay
-                std::set<size_t> node_ids = get_column_line_node_ids(bay_i - 1);
+                std::set<unsigned> node_ids = get_column_line_node_ids(bay_i - 1);
                 auto it = node_ids.begin();
-                for (size_t node_id_i = 0; node_id_i < node_ids.size() - 1; ++node_id_i) {
+                for (unsigned node_id_i = 0; node_id_i < node_ids.size() - 1; ++node_id_i) {
                     element_id++;
-                    elements_map.push_back(std::make_pair(element_id, std::vector<size_t>{*it, *(++it)}));
+                    elements_map.push_back(std::make_pair(element_id, std::vector<unsigned>{*it, *(++it)}));
                 }
                 // beams on each floor
                 for (int floor_i = 1; floor_i <= nfloors; ++floor_i) {
                     node_ids = get_beam_node_ids(bay_i, floor_i, true);
                     it = node_ids.begin();
-                    for (size_t node_id_i = 0; node_id_i < node_ids.size() - 1; ++node_id_i) {
+                    for (unsigned node_id_i = 0; node_id_i < node_ids.size() - 1; ++node_id_i) {
                         element_id++;
-                        elements_map.push_back(std::make_pair(element_id, std::vector<size_t>{*it, *(++it)}));
+                        elements_map.push_back(std::make_pair(element_id, std::vector<unsigned>{*it, *(++it)}));
                     }
                 }
             }
             // last column line
-            std::set<size_t> node_ids = get_column_line_node_ids(nbays);
+            std::set<unsigned> node_ids = get_column_line_node_ids(nbays);
             auto it = node_ids.begin();
-            for (size_t node_id_i = 0; node_id_i < node_ids.size() - 1; ++node_id_i) {
+            for (unsigned node_id_i = 0; node_id_i < node_ids.size() - 1; ++node_id_i) {
                 element_id++;
-                elements_map.push_back(std::make_pair(element_id, std::vector<size_t>{*it, *(++it)}));
+                elements_map.push_back(std::make_pair(element_id, std::vector<unsigned>{*it, *(++it)}));
             }
             return elements_map;
         }
@@ -372,14 +372,14 @@ class FrameMesh {
         /**
          * @brief Gets the IDs of the frame that will need to be restrained out-of-plane, which is all except bases.
          * 
-         * @return std::set<size_t> A set of node IDs representing the nodes that will need out-of-plane restrain.
+         * @return std::set<unsigned> A set of node IDs representing the nodes that will need out-of-plane restrain.
          */
-        std::set<size_t> get_out_of_plane_nodes()
+        std::set<unsigned> get_out_of_plane_nodes()
         {
-            std::set<size_t> column_bases = get_column_bases();
-            std::set<size_t> all_node_ids;
-            std::set<size_t> temp_id_set;
-            std::set<size_t> out_of_plane_nodes;
+            std::set<unsigned> column_bases = get_column_bases();
+            std::set<unsigned> all_node_ids;
+            std::set<unsigned> temp_id_set;
+            std::set<unsigned> out_of_plane_nodes;
             for (int floor_i = 1; floor_i <= nfloors; ++floor_i)
             {
                 temp_id_set = get_beam_line_node_ids(floor_i);
@@ -400,9 +400,9 @@ class FrameMesh {
         /**
          * @brief Get the node IDs that correspond to the bases of the column that will be fixed.
          * 
-         * @return std::set<size_t> nodes at the base of the column that will be fixed.
+         * @return std::set<unsigned> nodes at the base of the column that will be fixed.
          */
-        std::set<size_t> get_column_bases()
+        std::set<unsigned> get_column_bases()
         {
             return get_vertices_ids_at_floor(0);   
         }
@@ -430,9 +430,9 @@ class FrameMesh {
 /**
  * @brief prints out each element ID and its corresponding node IDs to the standard output stream.
  * 
- * @param element_map std::vector<std::pair<size_t, std::vector<size_t>>> that contains element ID followed by node IDs for this element.
+ * @param element_map std::vector<std::pair<unsigned, std::vector<unsigned>>> that contains element ID followed by node IDs for this element.
  */
-void inline read_element_map(std::vector<std::pair<size_t, std::vector<size_t>>> element_map)
+void inline read_element_map(std::vector<std::pair<unsigned, std::vector<unsigned>>> element_map)
 {
     for (auto& pair : element_map)
     {
@@ -443,9 +443,9 @@ void inline read_element_map(std::vector<std::pair<size_t, std::vector<size_t>>>
 /**
  * @brief prints out the node IDs and their xyz coordinates to the standard output stream.
  * 
- * @param nodes_coords_vector std::vector<std::pair<size_t, coords>> containing node IDs followed by coordinates of the node.
+ * @param nodes_coords_vector std::vector<std::pair<unsigned, coords>> containing node IDs followed by coordinates of the node.
  */
-void inline read_nodes_coords_vector(std::vector<std::pair<size_t, coords>> nodes_coords_vector)
+void inline read_nodes_coords_vector(std::vector<std::pair<unsigned, coords>> nodes_coords_vector)
 {
     for (auto& pair : nodes_coords_vector)
     {
